before final

+++ back/src/main/java/ru/andrew/hack/biv/service/CommonService.java
@@ -1,18 +1,15 @@
 package ru.andrew.hack.biv.service;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.springframework.stereotype.Service;
-import ru.andrew.hack.biv.dto.ObjectDto;
-import ru.andrew.hack.biv.dto.RiskDto;
-import ru.andrew.hack.biv.dto.TypeDto;
 import ru.andrew.hack.biv.model.ObjectDoc;
 import ru.andrew.hack.biv.model.Risk;
 import ru.andrew.hack.biv.model.Type;
 import ru.andrew.hack.biv.repos.ObjectRepository;
+import ru.andrew.hack.biv.repos.ProductRepository;
 import ru.andrew.hack.biv.repos.RiskRepository;
 import ru.andrew.hack.biv.repos.TypeRepository;
 
@@ -27,16 +24,16 @@ public class CommonService {
     private final ObjectRepository objectRepository;
     private final RiskRepository riskRepository;
     private final TypeRepository typeRepository;
-    private final ObjectMapper mapper;
+    private final ProductRepository productRepository;
 
     public Object get(String path) throws JSONException {
         String[] ids = path.split("\\.");
+        if (!productRepository.existsById(ids[0])) {
+            return "";
+        }
         switch (ids.length) {
             case 2 -> {
                 var objectDoc = objectRepository.findByPath(path);
-                log.info(objectDoc.toString());
-                //return mapper.convertValue(objectDoc.toString(), ObjectDto.class);
-                //return objectDoc.toString();
                 JSONObject jsonObject = new JSONObject();
                 jsonObject.put("id", objectDoc.getId());
                 jsonObject.put("name", objectDoc.getName());
@@ -47,18 +44,24 @@ public class CommonService {
             }
             case 3 -> {
                 var type = typeRepository.findByPath(path);
-                if (type == null) {
-                    return "";
-                }
-                //return mapper.convertValue(type.toString(), TypeDto.class);
+                JSONObject jsonObject = new JSONObject();
+                jsonObject.put("id", type.getId());
+                jsonObject.put("name", type.getName());
+                jsonObject.put("description", type.getDescription());
+                jsonObject.put("path", type.getPath());
+                jsonObject.put("parameters", type.getParameters());
+
                 return type.toString();
             }
             case 4 -> {
                 var risk = riskRepository.findByPath(path);
-                if (risk == null) {
-                    return "";
-                }
-                //return mapper.convertValue(risk.toString(), RiskDto.class);
+                JSONObject jsonObject = new JSONObject();
+                jsonObject.put("id", risk.getId());
+                jsonObject.put("name", risk.getName());
+                jsonObject.put("description", risk.getDescription());
+                jsonObject.put("path", risk.getPath());
+                jsonObject.put("parameters", risk.getParameters());
+
                 return risk.toString();
             }
             default -> {
@@ -69,26 +72,36 @@ public class CommonService {
 
     public Object post(String path, String body) throws JSONException {
         String[] ids = path.split("\\.");
+        if (!productRepository.existsById(ids[0])) {
+            return "";
+        }
         JSONObject jsonObj = new JSONObject(body);
         log.info(jsonObj.toString());
         switch (ids.length) {
             case 1, 2 -> {
                 var object = objectRepository.findByPath(path);
                 if (object == null) {
+                    log.info("null");
                     object = new ObjectDoc();
+                } else {
+                    log.info("not null");
                 }
                 object.setName(jsonObj.getString("name"));
                 object.setDescription(jsonObj.getString("description"));
-                jsonObj.remove("name");
-                jsonObj.remove("description");
-                object.setParameters(jsonObj.getJSONObject("parameters"));
+                object.setParameters(jsonObj.getString("parameters"));
                 if (ids.length == 1) {
                     objectRepository.save(object);
                     object.setPath(path + "." + object.getId());
+                    log.info(object.toString());
                     objectRepository.save(object);
                 } else {
-                    object.setPath(path);
+                    String id = object.getId();
+                    if (object.getId() == null) {
+                        return "";
+                    }
+                    object.setPath(path + "." + id);
                     objectRepository.save(object);
+                    log.info(objectRepository.findById(id).toString());
                 }
 
                 return object.getId();
@@ -100,10 +113,19 @@ public class CommonService {
                 }
                 type.setName(jsonObj.getString("name"));
                 type.setDescription(jsonObj.getString("description"));
-                type.setPath(path);
-                jsonObj.remove("name");
-                jsonObj.remove("description");
-                type.setParameters(jsonObj);
+                if (type.getPath() == null) {
+                    var newType = new Type();
+                    newType.setName(jsonObj.getString("name"));
+                    newType.setDescription(jsonObj.getString("description"));
+                    newType.setParameters(jsonObj.getString("parameters"));
+                    typeRepository.save(newType);
+                    newType.setPath(path + "." + newType.getId());
+                    typeRepository.save(newType);
+                    return newType.getId();
+                } else {
+                    type.setPath(path + "." + type.getId());
+                }
+                type.setParameters(jsonObj.getString("parameters"));
                 typeRepository.save(type);
                 return type.getId();
             }
@@ -114,10 +136,18 @@ public class CommonService {
                 }
                 risk.setName(jsonObj.getString("name"));
                 risk.setDescription(jsonObj.getString("description"));
+                if (risk.getPath() == null) {
+                    var newRisk = new ru.andrew.hack.biv.model.Risk();
+                    newRisk.setName(jsonObj.getString("name"));
+                    newRisk.setDescription(jsonObj.getString("description"));
+                    newRisk.setParameters(jsonObj.getString("parameters"));
+                    riskRepository.save(newRisk);
+                    newRisk.setPath(path + "." + newRisk.getId());
+                    riskRepository.save(newRisk);
+                    return newRisk.getId();
+                }
                 risk.setPath(path);
-                jsonObj.remove("name");
-                jsonObj.remove("description");
-                risk.setParameters(jsonObj);
+                risk.setParameters(jsonObj.getString("parameters"));
                 riskRepository.save(risk);
                 return risk.getId();
             }
@@ -129,6 +159,9 @@ public class CommonService {
 
     public String delete(String path) {
         String[] ids = path.split("\\.");
+        if (!productRepository.existsById(ids[0])) {
+            return "";
+        }
         switch (ids.length) {
             case 2 -> {
                 objectRepository.deleteByPath(path);
@@ -152,10 +185,54 @@ public class CommonService {
         List<ObjectDoc> objects = objectRepository.findAll();
         List<Type> types = typeRepository.findAll();
         List<Risk> risks = riskRepository.findAll();
-        var result = new ArrayList<Object>();
-        result.addAll(objects);
-        result.addAll(types);
-        result.addAll(risks);
+        var result = new ArrayList<>();
+        objects.forEach(obj-> {
+            JSONObject jsonObject = new JSONObject();
+            try {
+                jsonObject.put("id", obj.getId());
+                jsonObject.put("name", obj.getName());
+                jsonObject.put("description", obj.getDescription());
+                jsonObject.put("path", obj.getPath());
+                jsonObject.put("parameters", obj.getParameters());
+            } catch (JSONException e) {
+                throw new RuntimeException(e);
+            }
+            result.add(jsonObject);
+        });
+        types.forEach(obj-> {
+            JSONObject jsonObject = new JSONObject();
+            try {
+                jsonObject.put("id", obj.getId());
+                jsonObject.put("name", obj.getName());
+                jsonObject.put("description", obj.getDescription());
+                jsonObject.put("path", obj.getPath());
+                jsonObject.put("parameters", obj.getParameters());
+            } catch (JSONException e) {
+                throw new RuntimeException(e);
+            }
+            result.add(jsonObject);
+        });
+        risks.forEach(obj-> {
+            JSONObject jsonObject = new JSONObject();
+            try {
+                jsonObject.put("id", obj.getId());
+                jsonObject.put("name", obj.getName());
+                jsonObject.put("description", obj.getDescription());
+                jsonObject.put("path", obj.getPath());
+                jsonObject.put("parameters", obj.getParameters());
+            } catch (JSONException e) {
+                throw new RuntimeException(e);
+            }
+            result.add(jsonObject);
+        });
+
         return result;
+    }
+
+    public String deleteAll() {
+        objectRepository.deleteAll();
+        typeRepository.deleteAll();
+        riskRepository.deleteAll();
+        return "ok";
     }
 }
